[ETCM-126] split JsonRpcControllerSpec into JRCFixture JRCEthSpec JRCSpec
